refactor inputs date, format of date, add profile image, change first group exercises with start component
