--- src/components/Training/TrainingWiki/FilterExercises/MuscleGroupsList/MuscleGroupsList.module.scss
@@ -1,23 +1,0 @@
-.groupsLinks {
-  div {
-    width: 100%;
-    margin-bottom: 5px;
-    display: flex;
-    flex-direction: column;
-
-    &:hover {
-      background-color: rgba(121, 128, 211, 0.2);
-    }
-  }
-  a {
-    font-weight: 500;
-    cursor: pointer;
-    display: block;
-    padding: 5px;
-  }
-}
-
-.activeLink {
-  background-color: #e5e5ea;
-  border-radius: 8px;
-}
