+++ src/components/Users/Users.module.scss
@@ -10,11 +10,4 @@
       flex-direction: column;
     }
   }
-
-  span {
-    margin-top: 20px;
-    color: rgba(16, 104, 191, 0.55);
-    font-size: 21px;
-    text-shadow: rgba(87, 87, 190, 0.45) 1px 0 1px;
-  }
 }
